+++ apps/passport-server/src/services/issuanceService.ts
@@ -11,11 +11,11 @@ import {
 import { ArgumentTypeName, SerializedPCD } from "@pcd/pcd-types";
 import { RSAPCDPackage } from "@pcd/rsa-pcd";
 import {
+  getPublicKey,
+  getTicketData,
   ITicketData,
   RSATicketPCD,
-  RSATicketPCDPackage,
-  getPublicKey,
-  getTicketData
+  RSATicketPCDPackage
 } from "@pcd/rsa-ticket-pcd";
 import {
   SemaphoreSignaturePCD,
@@ -57,6 +57,7 @@ export class IssuanceService {
     const serialized = await Promise.all(
       pcds.map((pcd) => RSATicketPCDPackage.serialize(pcd))
     );
+
     return { pcds: serialized, folder: "Devconnect" };
   }
 
@@ -117,7 +118,7 @@ export class IssuanceService {
       };
     } catch (e) {
       logger("Error when consuming devconnect ticket", { error: e });
-      throw new Error("failed to check in");
+      throw new Error("failed to check in", {cause: e});
     }
   }