+++ apps/consumer-server/CHANGELOG.md
@@ -1,5 +1,12 @@
 # consumer-server
 
+## 0.0.13
+
+### Patch Changes
+
+- Updated dependencies
+  - @pcd/pcd-types@0.6.1
+
 ## 0.0.12
 
 ### Patch Changes