+++ apps/passport-server/CHANGELOG.md
@@ -1,5 +1,18 @@
 # passport-server
 
+## 0.0.17
+
+### Patch Changes
+
+- Updated dependencies
+  - @pcd/ethereum-ownership-pcd@0.6.1
+  - @pcd/passport-interface@0.6.1
+  - @pcd/pcd-types@0.6.1
+  - @pcd/rln-pcd@0.5.1
+  - @pcd/rsa-pcd@0.1.1
+  - @pcd/rsa-ticket-pcd@0.1.1
+  - @pcd/semaphore-group-pcd@0.6.1
+
 ## 0.0.16
 
 ### Patch Changes