+++ packages/webauthn-pcd/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @pcd/webauthn-pcd
 
+## 0.6.1
+
+### Patch Changes
+
+- update packages.
+- Updated dependencies
+  - @pcd/pcd-types@0.6.1
+
 ## 0.6.0
 
 ### Minor Changes